Fix date range picker demo from not switching between variants correctly
