+++ src/date-range-picker/use-date-range-picker.tsx
@@ -1,12 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { useEffect, useRef, useState } from 'react';
-import { DateRangePickerProps, Focusable } from './interfaces';
-import { ButtonProps } from '../button/interfaces';
-import useFocusVisible from '../internal/hooks/focus-visible';
-
-const VALID_RANGE: DateRangePickerProps.ValidRangeResult = { valid: true };
-
+import { useState } from 'react';
+import { DateRangePickerProps } from './interfaces';
+import { setTimeOffset } from './time-offset';
 /**
  * This function fills in a start and end time if they are missing.
  */
@@ -21,6 +17,23 @@ function fillMissingTime(value: DateRangePickerProps.AbsoluteValue | null) {
     startDate: startTime ? value.startDate : `${startDate}T00:00:00`,
     endDate: endTime ? value.endDate : `${endDate}T23:59:59`,
   };
+}
+
+export function formatValue(
+  value: null | DateRangePickerProps.Value,
+  { timeOffset, dateOnly }: { timeOffset: number; dateOnly: boolean }
+): null | DateRangePickerProps.Value {
+  if (!value || value.type === 'relative') {
+    return value;
+  }
+  if (dateOnly) {
+    return {
+      type: 'absolute',
+      startDate: value.startDate.split('T')[0],
+      endDate: value.endDate.split('T')[0],
+    };
+  }
+  return setTimeOffset(value, timeOffset);
 }
 
 function getDefaultMode(
@@ -40,18 +53,9 @@ function getDefaultMode(
   return relativeOptions.length > 0 ? 'relative' : 'absolute';
 }
 
-export function useDateRangePicker({
-  value,
-  relativeOptions,
-  rangeSelectorMode,
-  isValidRange,
-}: UseDateRangePickerProps) {
-  const focusVisible = useFocusVisible();
-  const scrollableContainerRef = useRef<HTMLDivElement | null>(null);
-  const applyButtonRef = useRef<ButtonProps.Ref>(null);
-
+export function useDateRangePicker({ value, relativeOptions, rangeSelectorMode }: UseDateRangePickerProps) {
   const [rangeSelectionMode, setRangeSelectionMode] = useState<'absolute' | 'relative'>(
-    getDefaultMode(value, relativeOptions, rangeSelectorMode!)
+    getDefaultMode(value, relativeOptions, rangeSelectorMode)
   );
 
   const [selectedAbsoluteRange, setSelectedAbsoluteRange] = useState<DateRangePickerProps.AbsoluteValue | null>(
@@ -62,52 +66,19 @@ export function useDateRangePicker({
     value?.type === 'relative' ? value : null
   );
 
-  const [applyClicked, setApplyClicked] = useState<boolean>(false);
-
-  const [validationResult, setValidationResult] = useState<
-    DateRangePickerProps.ValidRangeResult | DateRangePickerProps.InvalidRangeResult
-  >(VALID_RANGE);
-
-  useEffect(() => {
-    if (applyClicked) {
-      const visibleRange =
-        rangeSelectionMode === 'relative' ? selectedRelativeRange : fillMissingTime(selectedAbsoluteRange);
-
-      const newValidationResult = isValidRange(visibleRange);
-      setValidationResult(newValidationResult || VALID_RANGE);
-    }
-  }, [applyClicked, isValidRange, rangeSelectionMode, selectedRelativeRange, selectedAbsoluteRange]);
-
-  const focusRefs = {
-    default: useRef<Focusable>(null),
-    'absolute-only': useRef<Focusable>(null),
-    'relative-only': useRef<Focusable>(null),
-  };
-
-  useEffect(() => scrollableContainerRef.current?.focus(), [scrollableContainerRef]);
-
   return {
-    applyButtonRef,
-    focusVisible,
-    focusRefs,
-    scrollableContainerRef,
+    fillMissingTime,
     rangeSelectionMode,
     setRangeSelectionMode,
     selectedAbsoluteRange,
     setSelectedAbsoluteRange,
     selectedRelativeRange,
     setSelectedRelativeRange,
-    applyClicked,
-    setApplyClicked,
-    validationResult,
-    setValidationResult,
-    VALID_RANGE,
   };
 }
 
 export interface UseDateRangePickerProps {
   value: null | DateRangePickerProps.Value;
   relativeOptions: ReadonlyArray<DateRangePickerProps.RelativeOption>;
-  rangeSelectorMode?: DateRangePickerProps.RangeSelectorMode;
-  isValidRange: DateRangePickerProps.ValidationFunction;
+  rangeSelectorMode: DateRangePickerProps.RangeSelectorMode;
 }
